Write Recipe 기능 구현
Write Recipe 기능 구현

+++ src/main/java/kr/momjobgo/eyou/web/jpa/entity/TipEntity.java
@@ -20,5 +20,5 @@ public class TipEntity {
     private String text;
 
     @Column(name = "order_num")
-    private Byte orderNum;
+    private Short orderNum;
 }
